% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{DISCRETE RELEVANT LIFE COURSE MODEL}
\usage{
mediation(y, x, M, covariates = NULL, ...)
}
\arguments{
\item{y}{The output vector.}

\item{x}{The treatment variable.}

\item{M}{The mediator.}

\item{...}{Optional parameters to pass to the stan function.}

\item{Covariates}{A matrix of covariates.}
}
\value{
An list containing estimates for mediation.
}
\description{
Estimate Madathil's regression model
y_i = delta X_ij * weights_j + C_i' alpha + eps_i
}
