% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{MVmediation}
\alias{MVmediation}
\title{Multivariate Mediation}
\usage{
MVmediation(y, x, M, covariates = NULL, ...)
}
\arguments{
\item{y}{The output vector.}

\item{x}{The treatment variable.}

\item{M}{The mediator.}

\item{...}{Optional parameters to pass to the stan function.}

\item{Covariates}{A matrix of covariates.}
}
\value{
A rstan object. delta is the direct effect.
}
\description{
Multivariate Mediation
}
\examples{
n <- 200
z <- rnorm(n)
M <- z + rnorm(n)
delta <- c(2,1,3)
y <- sapply( delta, function(d) d * z + 2 * M + rnorm(n) )
}
