% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{DISCRETE RELEVANT LIFE COURSE MODEL}
\usage{
mediation(y, x, M, covariates = NULL, ...)
}
\arguments{
\item{y}{The output vector.}

\item{x}{The treatment variable.}

\item{M}{The mediator.}

\item{...}{Optional parameters to pass to the stan function.}

\item{Covariates}{A matrix of covariates.}
}
\value{
A rstan object. delta is the direct effect.
}
\description{
Estimate Madathil's regression model
y_i = delta X_ij * weights_j + C_i' alpha + eps_i
}
\examples{
n <- 200
z <- rnorm(n)
M <- z + rnorm(n)
y <- 2 * z + 2 * M + rnorm(n)
fit <- mediation(y, z, M, warmup = 2000, iter = 4000, chains = 4)
}
