% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{Mediation}
\usage{
mediation(y, x, M, covariates = NULL, ...)
}
\arguments{
\item{y}{The output vector.}

\item{x}{The treatment variable.}

\item{M}{The mediator.}

\item{...}{Optional parameters to pass to the stan function.}

\item{Covariates}{A matrix of covariates.}
}
\value{
A rstan object. delta is the direct effect.
}
\description{
Mediation
}
\examples{
n <- 200
p <- 3
z <- matrix( rnorm(n*p), ncol = p )
a <- c(1,2,3)
b <- c(0,2,1)
M <- c(z \%*\% a + rnorm(n))
delta <- c(2,1,3)
y <- sapply( 1:p, function(j) delta[j] * z[, j] + b[j] * M + rnorm(n) )
fit <- MVmediation(y, z, M, warmup = 2000, iter = 4000, chains = 4)
}
