% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{Mediation}
\usage{
mediation(y, x, M, covariates = NULL, ...)
}
\arguments{
\item{y}{The output vector.}

\item{x}{The treatment variable.}

\item{M}{The mediator.}

\item{...}{Optional parameters to pass to the stan function.}

\item{Covariates}{A matrix of covariates.}
}
\value{
A rstan object. delta is the direct effect.
}
\description{
Mediation
}
\examples{
n <- 200
z <- rnorm(n)
a <- 1
b <- 2
M <- z * a + rnorm(n)
delta <- 2
y <- delta * z + b * M + rnorm(n)
fit <- MVmediation(y, z, M, warmup = 2000, iter = 4000, chains = 4)
}
